rotate Meta Business Assets on E2E test runs
