+++ .github/workflows/product-creation-tests.yml
@@ -24,6 +24,7 @@ env:
   TEST_SITE_HOST: ${{ secrets.TEST_SITE_HOST }}
   WORDPRESS_PATH: '/tmp/wordpress'
   WP_DEBUG_LOG: '/tmp/wordpress/wp-content/debug.log'
+  WC_LOG_PATH: '/tmp/wordpress/wp-content/uploads/wc-logs/'
   PLUGIN_PATH: '/tmp/wordpress/wp-content/plugins/facebook-for-woocommerce'
   FB_E2E_TEST_COOKIE_NAME: 'facebook_test_id'
   FB_E2E_LOGGER_PATH: '/tests/e2e/lib/Logger.php'
@@ -61,14 +62,6 @@ env:
   # Error Whitelist (pipe-separated list of allowed non-critical errors)
   ERROR_WHITELIST: 'WC_Facebook_Google_Product_Category_Fields is not defined'
 
-  # Facebook Configuration
-  FB_EXTERNAL_BUSINESS_ID: ${{ secrets.FB_EXTERNAL_BUSINESS_ID }}
-  FB_ACCESS_TOKEN: ${{ secrets.FB_ACCESS_TOKEN }}
-  FB_PIXEL_ID: ${{ secrets.FB_PIXEL_ID }}
-  FB_PRODUCT_CATALOG_ID: ${{ secrets.FB_PRODUCT_CATALOG_ID }}
-  FB_PAGE_ID: ${{ secrets.FB_PAGE_ID }}
-  FB_BUSINESS_MANAGER_ID: ${{ secrets.FB_BUSINESS_MANAGER_ID }}
-
 jobs:
   e2e-tests:
     runs-on: ubuntu-latest
@@ -100,7 +93,11 @@ jobs:
         uses: shivammathur/setup-php@v2
         with:
           php-version: '8.1'
-          extensions: mysqli, zip, gd, curl, dom, imagick, fileinfo, mbstring
+          extensions: mysqli, zip, gd, curl, dom, fileinfo, mbstring
+          ini-values: |
+            date.timezone=UTC
+            memory_limit=512M
+            max_execution_time=300
 
       - name: Configure hostname
         run: |
@@ -326,6 +323,53 @@ jobs:
 
           echo "✅ PASS: No errors found in logs"
 
+      - name: Select Meta Business Assets
+        run: |
+          # Parse the JSON array of business assets
+          ASSETS='${{ secrets.FB_TEST_META_BUSINESS_ASSETS }}'
+
+          # Get the array length
+          ARRAY_LENGTH=$(echo "$ASSETS" | jq 'length')
+          echo "Total available business asset sets: $ARRAY_LENGTH"
+
+          # Use run_number as selector - it's sequential (1, 2, 3, ...) across all workflow runs
+          # This ensures each concurrent run gets a different business asset set
+          SELECTOR="${{ github.run_number }}"
+          echo "Using GitHub Run Number as selector: $SELECTOR"
+
+          # Calculate index using modulo
+          INDEX=$((SELECTOR % ARRAY_LENGTH))
+          echo "Selected asset set index: $INDEX"
+
+          # Extract the selected asset set
+          SELECTED_ASSET=$(echo "$ASSETS" | jq -c ".[$INDEX]")
+
+          # Extract values into variables
+          ACCESS_TOKEN=$(echo "$SELECTED_ASSET" | jq -r '.wc_facebook_access_token')
+          BUSINESS_MANAGER_ID=$(echo "$SELECTED_ASSET" | jq -r '.wc_facebook_business_manager_id')
+          EXTERNAL_BUSINESS_ID=$(echo "$SELECTED_ASSET" | jq -r '.wc_facebook_external_business_id')
+          PRODUCT_CATALOG_ID=$(echo "$SELECTED_ASSET" | jq -r '.wc_facebook_product_catalog_id')
+          PIXEL_ID=$(echo "$SELECTED_ASSET" | jq -r '.wc_facebook_pixel_id')
+          PAGE_ID=$(echo "$SELECTED_ASSET" | jq -r '.wc_facebook_page_id')
+
+          # Mask all sensitive values to prevent them from appearing in logs
+          echo "::add-mask::$ACCESS_TOKEN"
+          echo "::add-mask::$BUSINESS_MANAGER_ID"
+          echo "::add-mask::$EXTERNAL_BUSINESS_ID"
+          echo "::add-mask::$PRODUCT_CATALOG_ID"
+          echo "::add-mask::$PIXEL_ID"
+          echo "::add-mask::$PAGE_ID"
+
+          # Set environment variables for subsequent steps
+          echo "FB_ACCESS_TOKEN=$ACCESS_TOKEN" >> $GITHUB_ENV
+          echo "FB_BUSINESS_MANAGER_ID=$BUSINESS_MANAGER_ID" >> $GITHUB_ENV
+          echo "FB_EXTERNAL_BUSINESS_ID=$EXTERNAL_BUSINESS_ID" >> $GITHUB_ENV
+          echo "FB_PRODUCT_CATALOG_ID=$PRODUCT_CATALOG_ID" >> $GITHUB_ENV
+          echo "FB_PIXEL_ID=$PIXEL_ID" >> $GITHUB_ENV
+          echo "FB_PAGE_ID=$PAGE_ID" >> $GITHUB_ENV
+
+          echo "✅ Selected business asset set $INDEX for this run"
+
       - name: Install Facebook for WooCommerce plugin
         run: |
           cd ${{ env.WORDPRESS_PATH }}
@@ -347,13 +391,14 @@ jobs:
             cd ${{ env.WORDPRESS_PATH }}
           fi
 
-          # Configure Facebook connection before activation
-          wp option update wc_facebook_access_token "${{ secrets.FB_ACCESS_TOKEN }}" --allow-root
-          wp option update wc_facebook_merchant_access_token "${{ secrets.FB_ACCESS_TOKEN }}" --allow-root
-          wp option update wc_facebook_business_manager_id "${{ secrets.FB_BUSINESS_MANAGER_ID }}" --allow-root
-          wp option update wc_facebook_external_business_id "${{ secrets.FB_EXTERNAL_BUSINESS_ID }}" --allow-root
-          wp option update wc_facebook_product_catalog_id "${{ secrets.FB_PRODUCT_CATALOG_ID }}" --allow-root
-          wp option update wc_facebook_pixel_id "${{ secrets.FB_PIXEL_ID }}" --allow-root
+          # Configure Facebook connection before activation using selected business assets
+          wp option update wc_facebook_access_token "${{ env.FB_ACCESS_TOKEN }}" --allow-root
+          wp option update wc_facebook_merchant_access_token "${{ env.FB_ACCESS_TOKEN }}" --allow-root
+          wp option update wc_facebook_business_manager_id "${{ env.FB_BUSINESS_MANAGER_ID }}" --allow-root
+          wp option update wc_facebook_external_business_id "${{ env.FB_EXTERNAL_BUSINESS_ID }}" --allow-root
+          wp option update wc_facebook_product_catalog_id "${{ env.FB_PRODUCT_CATALOG_ID }}" --allow-root
+          wp option update wc_facebook_pixel_id "${{ env.FB_PIXEL_ID }}" --allow-root
+          wp option update wc_facebook_page_id "${{ env.FB_PAGE_ID }}" --allow-root
           wp option update wc_facebook_enable_server_to_server "yes" --allow-root
           wp option update wc_facebook_enable_pixel "yes" --allow-root
           wp option update wc_facebook_enable_advanced_matching "yes" --allow-root
@@ -362,7 +407,6 @@ jobs:
           wp option update wc_facebook_has_connected_fbe_2 "yes" --allow-root
           wp option update wc_facebook_has_authorized_pages_read_engagement "yes" --allow-root
           wp option update wc_facebook_enable_product_sync "yes" --allow-root
-          wp option update wc_facebook_page_id  "${{ secrets.FB_PAGE_ID }}" --allow-root
 
           # Activate the plugin (this triggers automatic sync)
           wp plugin activate facebook-for-woocommerce --allow-root
@@ -477,9 +521,6 @@ jobs:
           chmod 777 ${{ env.PLUGIN_PATH }}/tests/e2e/captured-events
 
       - name: Create test product for CAPI Pixel E2E tests
-        env:
-          WP_USERNAME: admin
-          WP_PASSWORD: admin
         run: |
           cd ${{ env.WORDPRESS_PATH }}
           # Create test product for ViewContent/AddToCart tests
@@ -508,15 +549,11 @@ jobs:
 
 
       - name: Run CAPI Pixel E2E tests
-        env:
-          TEST_PRODUCT_ID: ${{ env.TEST_PRODUCT_ID }}
         run: |
           cd ${{ env.PLUGIN_PATH }}
           xvfb-run --auto-servernum npx playwright test events-test*.spec.js --headed --project=chromium-wp-customer
 
       - name: Run Plugin level tests
-        env:
-          WC_LOG_PATH: ${{ env.WORDPRESS_PATH }}/wp-content/uploads/wc-logs
         run: |
           cd ${{ env.PLUGIN_PATH }}
           npx playwright test "plugin-.*\.spec\.js" --project=chromium-wp-admin
