+++ playwright.config.js
@@ -5,7 +5,7 @@ export default defineConfig({
   fullyParallel: true,
   forbidOnly: !!process.env.CI,
   retries: process.env.CI ? 2 : 0,
-  workers: process.env.CI ? 1 : undefined,
+  workers: process.env.CI ? 2 : 2,
   reporter: 'html',
   // Global test timeout - increased to 5 minutes for complex WordPress operations
   timeout: 1000000,
