+++ tests/e2e/product-creation.spec.js
@@ -232,167 +232,6 @@ test.describe('Facebook for WooCommerce - Product Creation E2E Tests', () => {
     }
   });
 
-  test('Test WordPress admin and Facebook plugin presence', async ({ page }, testInfo) => {
-
-    try {
-      // Navigate to plugins page with increased timeout
-      await page.goto(`${baseURL}/wp-admin/plugins.php`, {
-        waitUntil: 'domcontentloaded',
-        timeout: TIMEOUTS.MAX
-      });
-
-      // Check if Facebook plugin is listed
-      const pageContent = await page.content();
-      const hasFacebookPlugin = pageContent.includes('Facebook for WooCommerce') ||
-        pageContent.includes('facebook-for-woocommerce');
-
-      if (hasFacebookPlugin) {
-        console.log('✅ Facebook for WooCommerce plugin detected');
-      } else {
-        console.warn('⚠️ Facebook for WooCommerce plugin not found in plugins list');
-      }
-
-      // Verify no PHP errors
-      expect(pageContent).not.toContain('Fatal error');
-      expect(pageContent).not.toContain('Parse error');
-
-      console.log('✅ Plugin detection test completed');
-      logTestEnd(testInfo, true);
-
-    } catch (error) {
-      console.log(`⚠️ Plugin detection test failed: ${error.message}`);
-      logTestEnd(testInfo, false);
-      throw error;
-    }
-  });
-
-  test('Test basic WooCommerce product list', async ({ page }, testInfo) => {
-
-    try {
-      // Go to Products list with increased timeout
-      await page.goto(`${baseURL}/wp-admin/edit.php?post_type=product`, {
-        waitUntil: 'domcontentloaded',
-        timeout: TIMEOUTS.MAX
-      });
-
-      // Verify no PHP errors on products page
-      const pageContent = await page.content();
-      expect(pageContent).not.toContain('Fatal error');
-      expect(pageContent).not.toContain('Parse error');
-
-      // Check if WooCommerce is working
-      const hasProductsTable = await page.locator('.wp-list-table').isVisible({ timeout: TIMEOUTS.LONG });
-      if (hasProductsTable) {
-        console.log('✅ WooCommerce products page loaded successfully');
-      } else {
-        console.warn('⚠️ Products table not found');
-      }
-
-      console.log('✅ Product list test completed');
-      logTestEnd(testInfo, true);
-
-    } catch (error) {
-      console.log(`⚠️ Product list test failed: ${error.message}`);
-      logTestEnd(testInfo, false);
-      throw error;
-    }
-  });
-
-  test('Quick PHP error check across key pages', async ({ page }, testInfo) => {
-
-    try {
-      const pagesToCheck = [
-        { path: '/wp-admin/', name: 'Dashboard' },
-        { path: '/wp-admin/edit.php?post_type=product', name: 'Products' },
-        { path: '/wp-admin/plugins.php', name: 'Plugins' }
-      ];
-
-      for (const pageInfo of pagesToCheck) {
-        try {
-          console.log(`🔍 Checking ${pageInfo.name} page...`);
-          await page.goto(`${baseURL}${pageInfo.path}`, {
-            waitUntil: 'domcontentloaded',
-            timeout: TIMEOUTS.MAX
-          });
-
-          const pageContent = await page.content();
-
-          // Check for PHP errors
-          expect(pageContent).not.toContain('Fatal error');
-          expect(pageContent).not.toContain('Parse error');
-          // expect(pageContent).not.toContain('Warning: ');
-          // TODO: Do not dump the whole page content, just check for errors
-
-          // Verify admin content loaded
-          await page.locator('#wpcontent').isVisible({ timeout: TIMEOUTS.LONG });
-
-          console.log(`✅ ${pageInfo.name} page loaded without errors`);
-
-        } catch (error) {
-          console.log(`⚠️ ${pageInfo.name} page check failed: ${error.message}`);
-        }
-      }
-
-      logTestEnd(testInfo, true);
-    } catch (error) {
-      logTestEnd(testInfo, false);
-      throw error;
-    }
-  });
-
-  test('Test Facebook plugin deactivation and reactivation', async ({ page }, testInfo) => {
-
-    try {
-      // Navigate to plugins page
-      await page.goto(`${baseURL}/wp-admin/plugins.php`, {
-        waitUntil: 'domcontentloaded',
-        timeout: TIMEOUTS.MAX
-      });
-
-      // Look for Facebook plugin row
-      const pluginRow = page.locator('tr[data-slug="facebook-for-woocommerce"], tr:has-text("Facebook for WooCommerce")').first();
-
-      await pluginRow.waitFor({ state: 'visible', timeout: TIMEOUTS.LONG });
-      console.log('✅ Facebook plugin found');
-
-      // Check if plugin is currently active
-      const isActive = await pluginRow.locator('.active').isVisible({ timeout: TIMEOUTS.LONG });
-      const deactivateLink = pluginRow.locator('a:has-text("Deactivate")');
-      const reactivateLink = pluginRow.locator('a:has-text("Activate")');
-
-      if (isActive) {
-        console.log('Plugin is active, testing deactivation...');
-        await deactivateLink.waitFor({ state: 'visible', timeout: TIMEOUTS.LONG });
-        await deactivateLink.click();
-        await reactivateLink.waitFor({ state: 'visible', timeout: TIMEOUTS.LONG });
-        console.log('✅ Plugin deactivated');
-        await reactivateLink.click();
-        await deactivateLink.waitFor({ state: 'visible', timeout: TIMEOUTS.LONG });
-        console.log('✅ Plugin reactivated');
-      } else {
-        console.log('Plugin is inactive, testing activation...');
-        const activateLink = pluginRow.locator('a:has-text("Activate")');
-        await activateLink.waitFor({ state: 'visible', timeout: TIMEOUTS.LONG });
-        await activateLink.click();
-        await deactivateLink.waitFor({ state: 'visible', timeout: TIMEOUTS.LONG });
-        console.log('✅ Plugin activated');
-      }
-
-      // Verify no PHP errors after plugin operations
-      const pageContent = await page.content();
-      expect(pageContent).not.toContain('Fatal error');
-      expect(pageContent).not.toContain('Parse error');
-
-      console.log('✅ Plugin activation test completed');
-      logTestEnd(testInfo, true);
-
-    } catch (error) {
-      console.log(`⚠️ Plugin activation test failed: ${error.message}`);
-      logTestEnd(testInfo, false);
-      throw error;
-    }
-  });
-
   test('Create Composite product with WPC plugin', async ({ page }, testInfo) => {
     let simpleProductId = null;
     let variableProductId = null;
